update LB ports iff the provider is haproxy
The legacy neutron-lbaasv2 implementation based on haproxy is not aware
of security groupts at VIP creation and sets the default SG. The default
SG may not provide access from the cluster ingress, so we make sure in
this patch that the right SG is used.

This was not an issue with OpenStack's hybrid firewall because the
haproxy running on namespaces were bypassing the security groups. The
problem popped up when newer firewall implementations in Neutron fixed
that hole and exposed the wrong SG setting.

Finally, it is not possible to set the SG directly in Heat and that is a
wontfix bug there.

Signed-off-by: Antoni Segura Puimedon <antonisp@celebdor.com>
Change-Id: Ia5100609655fef2ad7906a2f9df63aa8e8d2f33f